# apps/docs/public/openapi-books.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Syncero Books API",
    "version": "1.0.0",
    "description": "API privada do **Syncero Books** — app para contadores.\n\n### Autenticação\n- **Rotas internas:** Bearer JWT emitido pelo Supabase (requer `Origin` de domínio Syncero)\n- **Rotas externas:** `x-api-key` (API Key do sistema) ou Bearer JWT\n\n### Restrição de Origem (Origin Guard)\nRotas internas verificam o header `Origin` em produção.\nApenas origens em `ALLOWED_ORIGINS` são aceitas — rotas externas não têm essa restrição.\n\n### Rotas externas (aceitam x-api-key)\n- `GET /api/transactions` — listar lançamentos de empresa externa\n- `POST /api/transactions` — criar lançamento para empresa externa\n- `POST /api/journal-entries` — classificar lançamento (criar partida dobrada)\n\n### Autorização interna\n- Empresas Flow: `accountant_companies.status = 'accepted'`\n- Empresas externas: `external_companies.accountant_id = userId`"
  },
  "servers": [
    {
      "url": "http://localhost:5175",
      "description": "Desenvolvimento local"
    },
    {
      "url": "https://syncero-books.vercel.app",
      "description": "Produção"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKey": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API Key gerada pelo sistema (GET/POST /api-keys). Aceita nas rotas externas como alternativa ao Bearer JWT."
      },
      "originGuard": {
        "type": "apiKey",
        "in": "header",
        "name": "Origin",
        "description": "Header de origem verificado nas rotas internas. Em produção, apenas origens Syncero são aceitas."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "ExternalCompany": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "accountant_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "cnpj": {
            "type": "string",
            "nullable": true
          },
          "trade_name": {
            "type": "string",
            "nullable": true
          },
          "tax_regime": {
            "type": "string",
            "enum": [
              "simples",
              "lucro_presumido",
              "lucro_real"
            ],
            "nullable": true
          },
          "segment": {
            "type": "string",
            "nullable": true,
            "enum": [
              "retail",
              "services",
              "manufacturing",
              "construction",
              "agribusiness",
              "healthcare",
              "education",
              "technology",
              "financial",
              "other"
            ]
          },
          "integration": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "key_prefix": {
            "type": "string",
            "example": "sk_live_abc12"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "ext_company_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "last_used_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "is_active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/companies": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Listar empresas Flow vinculadas",
        "description": "Retorna empresas onde o contador foi aceito (`accountant_companies.status = 'accepted'`).",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string"
                      },
                      "cnpj": {
                        "type": "string",
                        "nullable": true
                      },
                      "tax_regime": {
                        "type": "string",
                        "nullable": true
                      },
                      "segment": {
                        "type": "string",
                        "nullable": true
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado ou origem não autorizada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-access": "internal"
      }
    },
    "/api/companies/{id}": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Obter detalhes de empresa Flow",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "cnpj": {
                      "type": "string",
                      "nullable": true
                    },
                    "tax_regime": {
                      "type": "string",
                      "nullable": true
                    },
                    "segment": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem acesso"
          }
        },
        "x-access": "internal"
      }
    },
...
    "/api/external-companies": {
      "get": {
        "tags": [
          "external-companies"
        ],
        "summary": "Listar empresas externas do contador",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExternalCompany"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado ou origem não autorizada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-access": "internal"
      },
      "post": {
        "tags": [
          "external-companies"
        ],
        "summary": "Criar empresa externa",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "cnpj": {
                    "type": "string",
                    "nullable": true
                  },
                  "trade_name": {
                    "type": "string",
                    "nullable": true
                  },
                  "tax_regime": {
                    "type": "string",
                    "enum": [
                      "simples",
                      "lucro_presumido",
                      "lucro_real"
                    ],
                    "nullable": true
                  },
                  "segment": {
                    "type": "string",
                    "nullable": true,
                    "enum": [
                      "retail",
                      "services",
                      "manufacturing",
                      "construction",
                      "agribusiness",
                      "healthcare",
                      "education",
                      "technology",
                      "financial",
                      "other"
                    ]
                  },
                  "notes": {
                    "type": "string",
                    "nullable": true
                  },
                  "seed_plan": {
                    "type": "boolean",
                    "default": false,
                    "description": "Se true, cria o Plano de Contas padrão CFC automaticamente"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Empresa criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalCompany"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado ou origem não autorizada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-access": "internal"
      }
    },
    "/api/external-companies/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "tags": [
          "external-companies"
        ],
        "summary": "Obter empresa externa",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalCompany"
                }
              }
            }
          },
          "403": {
            "description": "Sem acesso"
          }
        },
        "x-access": "internal"
      },
      "patch": {
        "tags": [
          "external-companies"
        ],
        "summary": "Atualizar empresa externa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Campos editáveis. **accountant_id é imutável** — não pode ser alterado após criação.",
                "x-non-writable": [
                  "accountant_id",
                  "id",
                  "created_at",
                  "updated_at"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "cnpj": {
                    "type": "string",
                    "nullable": true
                  },
                  "trade_name": {
                    "type": "string",
                    "nullable": true
                  },
                  "tax_regime": {
                    "type": "string",
                    "enum": [
                      "simples",
                      "lucro_presumido",
                      "lucro_real"
                    ],
                    "nullable": true
                  },
                  "segment": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalCompany"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado ou origem não autorizada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-access": "internal"
      }
    },
...
    "/api/api-keys": {
      "get": {
        "tags": [
          "api-keys"
        ],
        "summary": "Listar chaves de API",
        "parameters": [
          {
            "name": "company_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "ext_company_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKey"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado ou origem não autorizada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-access": "internal"
      },
      "post": {
        "tags": [
          "api-keys"
        ],
        "summary": "Criar chave de API",
        "description": "O token completo é retornado **uma única vez** no campo `key`. Armazene-o imediatamente — não é possível recuperá-lo depois.\n\nVerifica que o contador tem acesso à empresa informada antes de criar a chave.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "company_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "ext_company_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiKey"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string",
                          "description": "Token completo — exibido somente nesta resposta",
                          "example": "sk_live_abc123xyz..."
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado ou origem não autorizada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-access": "internal"
      }
    },
    "/api/api-keys/{id}/revoke": {
      "patch": {
        "tags": [
          "api-keys"
        ],
        "summary": "Revogar chave de API",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado ou origem não autorizada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-access": "internal"
      }
    },
...
  }
}